MAINT: stats: rewrite find_repeats
* No more Fortran extension for something we can do with NumPy
* Refactored to remove duplicate implementation
* Document the implicit float64 cast (no longer necessary...)
