+++ scipy/stats/bento.info
@@ -8,8 +8,6 @@ Library:
             statlib/swilk.f
     Extension: statlib
         Sources: statlib.pyf
-    Extension: futil
-        Sources: futil.f
     Extension: mvn
         Sources: mvn.pyf, mvndst.f
     Extension: vonmises_cython
