+++ scipy/stats/setup.py
@@ -31,11 +31,6 @@ def configuration(parent_package='',top_path=None):
         sources=['_rank.c'],          # FIXME: use cython source
     )
 
-    # add futil module
-    config.add_extension('futil',
-        sources=['futil.f'],
-    )
-
     # add mvn module
     config.add_extension('mvn',
         sources=['mvn.pyf','mvndst.f'],
